[fix] fix type of BarStackSegment.Value. type of BarStackSegment.Value should be float, because we should tailor to the Bar.Value type.

+++ HoloGraphLibrary/src/com/echo/holographlibrary/BarGraph.java
@@ -139,7 +139,7 @@ public class BarGraph extends View {
 
                 if(p.getStackedBar()){
                     ArrayList<BarStackSegment> values = new ArrayList<BarStackSegment>(p.getStackedValues());
-                    int prevValue = 0;
+                    float prevValue = 0;
                     for(BarStackSegment value : values) {
                         value.Value += prevValue;
                         prevValue += value.Value;
